Random trajectories

+++ DuckHunt/utils.h
@@ -43,7 +43,7 @@ struct Button
 struct Duck
 {
     Sprite *sprite;
-    int speed;      //vitesse du canard
+    float speed;      //vitesse du canard
     int nbr_sprite; //nbr d'images qui va composer l'animation
     int mvt_x;
     int mvt_y;
